requirements.txt
numpy>=1.19.5
opencv-python>=4.5.0
pillow>=8.0.0
tensorflow>=2.5.0
librosa>=0.8.0
transformers>=4.5.0
torch>=1.8.0
torchvision>=0.9.0
sounddevice>=0.4.0
scikit-learn>=0.24.0
pandas>=1.2.0
tqdm>=4.60.0